Make int instead of long for scala map api

+++ october.thrift
@@ -121,7 +121,7 @@ service Recommender {
      * @param user_id, the user to query for
      * @param limit, the maximum amount of tokens to return 
      */
-    map<string, i64> user_top_terms(1: required i64 user_id, 2: required i64 limit) throws (1: NotFoundException nfe),
+    map<string, i64> user_top_terms(1: required i64 user_id, 2: required i32 limit) throws (1: NotFoundException nfe),
 
     /** Return a list of documents in sorted order of relevance for a search query
      * @param query, a map of tokens to their weight
